% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_floating_axes.R
\name{add_floating_y}
\alias{add_floating_y}
\title{Define broken axis theme}
\usage{
add_floating_y(x)
}
\arguments{
\item{vector}{with (min, max) values for floating y-axis}
}
\description{
\code{add_floating_y} makes the included CSIRO font available in R. This process
only needs to be completed once.
}
