% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_logo.R
\name{add_logo}
\alias{add_logo}
\title{Return function to add a logo to a figure}
\usage{
add_logo(p, x = 0.07, y = 0.03, w = 0.3, u = "inches",
  logo = "csiro")
}
\arguments{
\item{p}{ggplot object}

\item{x}{float indicating the x coordinate of the logo}

\item{y}{float indicating the  ycoordinate of the logo}

\item{w}{float indicating the width of the logo}

\item{u}{String indicating the units of the logo width: either inches or cm}

\item{logo}{String indicating which logo to use. There are two options: csiro, which is the usual CSIRO logo; data61, which is the CSIRO logo plus the Data61.}
}
\description{
Return function to add a logo to a figure
}
