% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_ggbluebadge.R
\name{csiro_cols}
\alias{csiro_cols}
\title{Function to extract csiro colors as hex codes}
\usage{
csiro_cols(...)
}
\arguments{
\item{...}{Character names of csiro_colors}
}
\description{
Function to extract csiro colors as hex codes
}
