% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_ggcsiro.R
\name{csiro_pal}
\alias{csiro_pal}
\title{Return function to interpolate a csiro color palette}
\usage{
csiro_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in csiro_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a csiro color palette
}
