% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_ggcsiro.R
\name{scale_fill_csiro}
\alias{scale_fill_csiro}
\title{Fill scale constructor for csiro colors}
\usage{
scale_fill_csiro(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in drsimonj_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}
}
\description{
Fill scale constructor for csiro colors
}
