% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_ggbluebadge.R
\name{show_my_palette}
\alias{show_my_palette}
\title{Display palette of interest}
\usage{
show_my_palette(palette = "main")
}
\arguments{
\item{palette}{Character name of palette in drsimonj_palettes}
}
\description{
Display palette of interest
}
