% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_csiro.R, R/theme_csiro_dark.R,
%   R/theme_csiro_darkta61.R, R/theme_csiro_data61.R
\name{theme_csiro}
\alias{theme_csiro}
\alias{theme_csiro_dark}
\alias{theme_csiro_darkta61}
\alias{theme_csiro_data61}
\title{Minimal Theme}
\usage{
theme_csiro(base_family = "sans", base_size = 12,
  base_color = "black", line_size = 0.2, baseline = FALSE,
  baseline.x = baseline, baseline.y = baseline,
  baseline.size = line_size * 3, background.color = "white",
  panel.border = TRUE, panel.border.size = line_size * 2,
  legend.box = FALSE, grid = FALSE, grid.minor = grid,
  grid.x = grid, grid.y = grid, grid.minor.x = grid.minor,
  grid.minor.y = grid.minor, grid.color = "black", ticks = TRUE,
  ticks.x = ticks, ticks.y = ticks, title.family = "sans",
  title.face = "bold", title.color = csiro_cols("midnight blue"),
  subtitle.family = base_family, subtitle.face = "plain",
  subtitle.color = csiro_cols("midday blue"),
  caption.family = base_family, caption.face = "plain",
  caption.color = "black", axis.title.family = base_family,
  axis.title.face = "plain", axis.title.color = base_color,
  axis.text.family = base_family, axis.text.face = "plain",
  axis.text.color = "black", legend.title.family = axis.title.family,
  legend.title.face = axis.title.face,
  legend.title.color = axis.title.color,
  legend.text.family = axis.text.family,
  legend.text.face = axis.text.face,
  legend.text.color = axis.text.color, aspect.ratio = NULL)

theme_csiro_dark(base_family = "sans", base_size = 12,
  base_color = "white", line_size = 0.2, baseline = FALSE,
  baseline.x = baseline, baseline.y = baseline,
  baseline.size = line_size * 3, background.color = "white",
  panel.border = TRUE, panel.border.size = line_size * 2,
  legend.box = FALSE, grid = FALSE, grid.minor = grid,
  grid.x = grid, grid.y = grid, grid.minor.x = grid.minor,
  grid.minor.y = grid.minor, grid.color = "white", ticks = TRUE,
  ticks.x = ticks, ticks.y = ticks, title.family = "sans",
  title.face = "bold", title.color = csiro_cols("midday blue"),
  subtitle.family = base_family, subtitle.face = "plain",
  subtitle.color = base_color, caption.family = base_family,
  caption.face = "italic", caption.color = "white",
  axis.title.family = base_family, axis.title.face = "plain",
  axis.title.color = base_color, axis.text.family = base_family,
  axis.text.face = "plain", axis.text.color = "white",
  legend.title.family = axis.title.family,
  legend.title.face = axis.title.face,
  legend.title.color = axis.title.color,
  legend.text.family = axis.text.family,
  legend.text.face = axis.text.face,
  legend.text.color = axis.text.color, aspect.ratio = NULL)

theme_csiro_darkta61(base_family = "Arial", base_size = 12,
  base_color = "black", line_size = 0.2, baseline = FALSE,
  baseline.x = baseline, baseline.y = baseline,
  baseline.size = line_size * 3, background.color = "white",
  panel.border = TRUE, panel.border.size = line_size * 2,
  legend.box = FALSE, grid = FALSE, grid.minor = grid,
  grid.x = grid, grid.y = grid, grid.minor.x = grid.minor,
  grid.minor.y = grid.minor, grid.color = "black", ticks = TRUE,
  ticks.x = ticks, ticks.y = ticks, title.family = "Arial",
  title.face = "bold", title.color = csiro_cols("mint"),
  subtitle.family = base_family, subtitle.face = "italic",
  subtitle.color = base_color, caption.family = base_family,
  caption.face = "italic", caption.color = "black",
  axis.title.family = base_family, axis.title.face = "plain",
  axis.title.color = base_color, axis.text.family = base_family,
  axis.text.face = "plain", axis.text.color = "black",
  legend.title.family = axis.title.family,
  legend.title.face = axis.title.face,
  legend.title.color = axis.title.color,
  legend.text.family = axis.text.family,
  legend.text.face = axis.text.face,
  legend.text.color = axis.text.color, aspect.ratio = NULL)

theme_csiro_data61(base_family = "Arial", base_size = 12,
  base_color = "white", line_size = 0.2, baseline = FALSE,
  baseline.x = baseline, baseline.y = baseline,
  baseline.size = line_size * 3, background.color = "white",
  panel.border = TRUE, panel.border.size = line_size * 2,
  legend.box = FALSE, grid = FALSE, grid.minor = grid,
  grid.x = grid, grid.y = grid, grid.minor.x = grid.minor,
  grid.minor.y = grid.minor, grid.color = "grey30", ticks = TRUE,
  ticks.x = ticks, ticks.y = ticks, title.family = "Arial",
  title.face = "bold", title.color = csiro_cols("light mint"),
  subtitle.family = base_family, subtitle.face = "italic",
  subtitle.color = base_color, caption.family = base_family,
  caption.face = "italic", caption.color = "white",
  axis.title.family = base_family, axis.title.face = "plain",
  axis.title.color = base_color, axis.text.family = base_family,
  axis.text.face = "plain", axis.text.color = "white",
  legend.title.family = axis.title.family,
  legend.title.face = axis.title.face,
  legend.title.color = axis.title.color,
  legend.text.family = axis.text.family,
  legend.text.face = axis.text.face,
  legend.text.color = axis.text.color, aspect.ratio = NULL)
}
\arguments{
\item{base_family}{Base font family (default: \code{sans})}

\item{base_size}{Base font size (default: \code{12} pt)}

\item{base_color}{Base font color (default: \code{grey30})}

\item{line_size}{Base line size for grid, panel (default: \code{0.2})}

\item{baseline}{Whether or not to show baselines on axes (default: \code{FALSE})}

\item{baseline.x}{Whether or not to show baselines on X axis (default: same as \code{baseline})}

\item{baseline.y}{Whether or not to show baselines on X axis (default: same as \code{baseline})}

\item{baseline.size}{Size of axis baseline (default: \code{line_size * 3})}

\item{background.color}{Background color (default: \code{white})}

\item{panel.border}{Whether or not to draw a border around panels (default: \code{FALSE})}

\item{panel.border.size}{Size of panel border (default: \code{line_size * 2})}

\item{legend.box}{Whether or not to draw box around legend (default: \code{TRUE})}

\item{grid}{Whether or not to show grid lines (default: \code{TRUE})}

\item{grid.minor}{Whether or not to show minor grid lines (default: same as \code{grid})}

\item{grid.x}{Whether or not to show grid lines on X axis (default: same as \code{grid})}

\item{grid.y}{Whether or not to show grid lines on Y axis (default: same as \code{grid})}

\item{grid.minor.x}{Whether or not to show minor grid lines on X axis (default: same as \code{grid.minor})}

\item{grid.minor.y}{Whether or not to show minor grid lines on Y axis (default: same as \code{grid.minor})}

\item{grid.color}{Color for the grid, ticks, baseline, and legend box (default: \code{grey80})}

\item{ticks}{Whether or not to show axis tick marks (default: \code{FALSE})}

\item{ticks.x}{Whether or not to show X axis tick marks (default: same as \code{ticks})}

\item{ticks.y}{Whether or not to show Y axis tick marks (default: same as \code{ticks})}

\item{title.family}{Font family for plot title (default: \code{"sans"})}

\item{title.face}{Font face for plot title (default: \code{bold})}

\item{title.color}{Color of plot title (default: \code{base_color})}

\item{subtitle.family}{Font family for plot subtitle (default: \code{base_family})}

\item{subtitle.face}{Font face for plot subtitle (default: \code{italic})}

\item{subtitle.color}{Color of plot subtitle (default: \code{base_color})}

\item{caption.family}{Font family for plot caption (default: \code{base_family})}

\item{caption.face}{Font face for plot caption (default: \code{"italic"})}

\item{caption.color}{Color of plot title (default: \code{grey70})}

\item{axis.title.family}{Font family for axis titles (default: \code{base_family})}

\item{axis.title.face}{Font face for axis titles (default: \code{"plain"})}

\item{axis.title.color}{Color of axis titles (default: \code{base_color})}

\item{axis.text.family}{Font family for axis text (default: \code{base_family})}

\item{axis.text.face}{Font face for axis text (default: \code{plain})}

\item{axis.text.color}{Color of axis text (default: \code{grey50})}

\item{legend.title.family}{Font family for legend title (default: \code{axis.title.family})}

\item{legend.title.face}{Font face for legend title (default: \code{axis.title.face})}

\item{legend.title.color}{Color of legend title (default: \code{axis.title.color})}

\item{legend.text.family}{Font family for legend text (default: \code{base_family})}

\item{legend.text.face}{Font face for legend title (default: \code{axis.text.face})}

\item{legend.text.color}{Color of legend text (default: \code{axis.text.color})}

\item{aspect.ratio}{Aspect ratio for panels (no default)}
}
\value{
A named list containing ggplot2 theme information
}
\description{
\code{theme_csiro} is a minimal yet flexible \link{ggplot2} theme designed to put
the data front and center. It uses the sans font.
}
\section{Installing and Using Lato}{

Before using this theme, you must

extrafont::font_import()
extrafont::loadfonts()
}

\examples{
\dontrun{
library(ggplot2)
library(ggbluebadge)

ggplot(mtcars, aes(wt, mpg, color = as.factor(cyl) )) +
    facet_grid(cyl ~ .) +
    geom_line() +
    geom_point() +
    scale_color_hue(name = "# Cylinders") +
    labs(
        x = "Weight (tons)",
        y = "Fuel Efficiency (mpg)",
        title = "Just Buy a Convertible",
        subtitle = "Lighter cars are more efficient. Shave off some weight by\\nditching the roof, and drive your savings to the bank!",
        caption = "Source: Motor Trend"
    ) +
    theme_csiro()
}
}
