% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_floating_axes.R
\name{theme_floating_axes}
\alias{theme_floating_axes}
\title{Define broken axis theme}
\usage{
theme_floating_axes(...)
}
\description{
\code{theme_floating_axes} makes the included CSIRO font available in R. This process
only needs to be completed once.
}
